adding comments to controllers and modifying foreign ids types

+++ models/article.go
@@ -6,7 +6,7 @@ import (
 
 type Article struct {
 	gorm.Model
-	UserId   uint
+	UserId   uint64
 	User     User
 	Title    string
 	Content  string
